Abstracting common part for actions
