+++ components/nexweave/actions/create-campaign-experience/create-campaign-experience.mjs
@@ -1,4 +1,4 @@
-import nexweave from "../../nexweave.app.mjs";
+import common from "../common.mjs";
 
 export default {
   key: "nexweave-create-campaign-experience",
@@ -6,45 +6,13 @@ export default {
   description: "Generates a campaign experience based on a selected campaign. [See the documentation](https://documentation.nexweave.com/nexweave-api)",
   version: "0.0.1",
   type: "action",
-  props: {
-    nexweave,
-    campaignId: {
-      propDefinition: [
-        nexweave,
-        "campaignId",
-      ],
-      reloadProps: true,
+  ...common,
+  methods: {
+    getSummary() {
+      return "Successfully created campaign experience";
     },
-  },
-  async additionalProps() {
-    const { campaignId } = this;
-    if (!campaignId) return {};
-
-    const campaignDetails = await this.nexweave.getCampaignDetails(campaignId);
-    return Object.fromEntries(campaignDetails?.result?.variables?.map?.((variable) => ([
-      variable.key,
-      {
-        type: "string",
-        label: `Variable: "${variable.key}"`,
-        description: `Default value: "${variable.default}"`,
-        optional: true,
-      },
-    ])) ?? {});
-  },
-  async run({ $ }) {
-    const {
-      nexweave, campaignId, ...data
-    } = this;
-
-    const response = await nexweave.createCampaignExperience({
-      $,
-      data: {
-        campaign_id: campaignId,
-        data,
-      },
-    });
-
-    $.export("$summary", "Successfully created campaign experience");
-    return response;
+    async createExperience(args) {
+      return this.nexweave.createCampaignExperience(args);
+    },
   },
 };
